add folder dtos
