+++ src/dtos/employee-create.dto.ts
@@ -1,0 +1,15 @@
+import { IsString, IsNotEmpty, IsNumber } from 'class-validator';
+
+export class CreateEmployeeDto {
+  @IsString()
+  @IsNotEmpty()
+  name: string;
+
+  @IsString()
+  @IsNotEmpty()
+  position: string;
+
+  @IsNumber()
+  @IsNotEmpty()
+  salary: number;
+}
