+++ src/dtos/employee-update.dto.ts
@@ -1,0 +1,15 @@
+import { IsString, IsOptional, IsNumber } from 'class-validator';
+
+export class UpdateEmployeeDto {
+  @IsString()
+  @IsOptional()
+  name?: string;
+
+  @IsString()
+  @IsOptional()
+  position?: string;
+
+  @IsNumber()
+  @IsOptional()
+  salary?: number;
+}
