--- src/employees/employee.dto.ts
@@ -1,11 +1,0 @@
-export class CreateEmployeeDto {
-    name: string;
-    position: string;
-    salary: number;
-  }
-  
-  export class UpdateEmployeeDto {
-    name?: string;
-    position?: string;
-    salary?: number;
-  }
